app/pyproject.toml
[tool.poetry]
name = "procurador"
version = "0.1.0"
description = ""
authors = ["Rhuan Barros <rhuanbarros@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
unstructured = {extras = ["all-docs"], version = "^0.14.0"}
langchain = "^0.2.0"
langchain-weaviate = "^0.0.2"
langchain-nomic = "^0.1.0"
langchain-community = "^0.2.0"
gpt4all = "^2.6.0"
tiktoken = "^0.7.0"
langgraph = "^0.0.50"
tavily-python = "^0.3.3"
weaviate-client = "^4.6.2"
fastapi-injector = "^0.5.4"
langchain-text-splitters = "^0.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
